requirements.txt
streamlit
tensorflow==2.15.0
tensorflow-hub
pillow
numpy<2
